Struct atomic_store::atomic_store::AtomicStore

pub struct AtomicStore {
    file_path: PathBuf,
    file_pattern: String,
    file_counter: u32,
    last_counter: Option<u32>,
    resources: HashMap<String, Arc<RwLock<VersionSyncHandle>>>,
    commit_timeout: Duration,
}

The central index of an atomic version of truth across multiple persisted data structures; Guarantees that all managed resources can be loaded in a consistent state across an entire logical entity.

Fields
file_path: PathBuffile_pattern: Stringfile_counter: u32last_counter: Option<u32>resources: HashMap<String, Arc<RwLock<VersionSyncHandle>>>commit_timeout: Duration
Implementations

impl AtomicStore

pub fn open(load_info: AtomicStoreLoader) -> Result<AtomicStore>

pub fn set_commit_timeout(&mut self, timeout: Duration)

Set the commit timeout. By default, AtomicStore will wait 100 milliseconds for all logs to be committed.

pub fn commit_version(&mut self) -> Result<()>

Commit the version. Note that all logs and stores must call .commit_version() or .skip_version() before this function is called.

This will timeout after 100 milliseconds (configurable with set_commit_timeout). If you want to disable this timeout, set the ATOMIC_STORE_NO_TIMEOUT environment variable before calling AtomicStore::open.