Module atomic_store::fixed_append_log

Structs

FixedAppendLog

For now, this is implemented with a direct copy from memory to file, using native order, but the order is recorded in a header, and can be used to support transfer in the future.

IndexContents

Iter

Constants

BYTE_DISORDER

BYTE_ORDER

Functions

compute_location

format_backup_index_file_path

format_index_file_path

format_range_file_path

format_working_index_file_path

load_existing_index